requirements.txt
atproto>=0.0.56,<1
requests>=2.31.0,<3
colorama>=0.4.6,<1
python-dotenv>=1.0.1,<2
